% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{ty_mst_default}
\alias{ty_mst_default}
\title{A default 1-3 MST: 12 routing items, easy/medium/hard modules of 24}
\usage{
ty_mst_default(center = 0)
}
\arguments{
\item{center}{Center of the difficulty range.}
}
\value{
A `ty_mst`.
}
\description{
A default 1-3 MST: 12 routing items, easy/medium/hard modules of 24
}
\examples{
ty_mst_default()$cuts
}
